modularising simulationtools role
simulationtools has split into

 - simulationtools: what used to be 'dependencies.yml'
 - quantum-espresso: what used to be 'quantumespresso.yml'
 - aiida: what used to be 'aiida-*.yml'

The roles are meant to be executed in this order.
The aiida role will recognise, whether the quantum-espresso role
has run and install the codes it provides.

Note: This was a significant restructuring effort. The ansible script
runs through, but some more testing is certainly required.