add way of compacting VM disks
simply deleting stuff from the VM disk does not result in a reduction
of size - for this one needs to compact the image, and this is
currently not possible in the vmdk format we are using.

The ./compact_hd.sh script takes care of the rather tedious procedure
of actually compacting the disk.