decouple aiida plugins from presence of code
fix #65

The installation of plugins is no longer coupled to the local
presence of a code (only the installation of the local code is).

The aiida role now has a new variable `aiida_plugin_versions`
which can be used to specify which plugins should be installed
(be default all are).